Refuse to join different sides of a network partition back together.
If we allowed this, then the registry would self-implode.

Eg, there are two nodes that are separated by a network partition. Each
node will remove the pids that are present on the other node from the
CRDT. If these nodes are allowed to reconnect, then they will remove
each other's nodes from the registry, leaving it empty (except for any
new pids added during the partition). To keep things safe if and until
we figure out a solution to this problem, it's impossible to join hordes
that have each experienced a neighbour disappearing uncleanly.