fixing bug where dynamic supervisor would always remove child
REASON: This bug was caused because the CRDT mutation occured in delete_child
function of the Horde.DynamicSupervisor. On toplogy changes duplicate
children needed to be removed from one of the existing supervisor, and
a call was made to terminate the child which would inadvrtently mutate
the CRDT to remove the child. However, the correct behavior is to just
stop the child.

FIX: The fix was to be more explicit where we mutate the CRDT in Horde.DynamicSupervisor,
moving it explictly to cases where either the process is temporary or transient with
a normal shutdown case.