Added method for controlling rebalance during operation.
Added another config for process_redistribution, :manual.
When configured for manual, an element in the supervisor's CRDT is
updated so that all supervisors in the cluster can detect a change in the flag.
The diff populates the state which is then checked when a diff occurs
and when node membership changes.