Fix crashes related to missing child ids in lookup table
I noticed that when a cluster is restarting, Horde tends to crash
because it does some administration that is incorrect. The child
lookup for the linked ETS table can fail but this was not handled on
all places in the Dynamic Supervisor implementation.

This fix addresses three places that resulted in crashes because `nil`
was not handled from the get_item / pop_item calls.

Note that there were several other places where this was already
addressed.