Fix crash when $gen_restart message is received for wrong child state
ProcessesSupervisor crashes on the $gen_restart message when a pid of
a child is given that is known to the supervisor but not in the
restarting state. This should of course never happen but in the case
it does, it is better to ignore the message than to crash the
supervisor.