Fixing version compatibility during upgrade
As indicated to me by a colleague after a code review, the implementation will cause issues when upgrading from the current version to the new version. The tuple size for :proxy_operation changed from 3 to 4, which means nodes with old and new versions won't be able to proxy to eachother.

This commit mitigates that to a degree. The new implementation supports both tuple size 3 and 4. And when :proxy_message_ttl is set to :infinity (the default), `proxy_to_node` will use tuple size 3. This means that upgrade is possible if during the upgrade `:proxy_message_ttl` option is not set, or set to :infinity. If it is set to an integer value, the upgrade will cause proxy messeage from upgraded nodes to old nodes to fail. The CHANGELOG would need to reflect this as an upgrade risk.